Fix parse_unit(). to_i -> to_f

+++ lib/fluent/plugin/in_top_parser.rb
@@ -63,11 +63,11 @@ module Fluent
       begin 
         case s[-1]
         when "m"
-          s[0, s.length-1].to_i * 1024
+          s[0, s.length-1].to_f * 1024
         when "g"
-          s[0, s.length-1].to_i * 1024 * 1024
+          s[0, s.length-1].to_f * 1024 * 1024
         when "t"
-          s[0, s.length-1].to_i * 1024 * 1024 * 1024
+          s[0, s.length-1].to_f * 1024 * 1024 * 1024
         else
           s.to_i
         end
